Backend/requirements.txt
google-generativeai==0.7.2
moviepy==1.0.3
Pillow==10.4.0
requests==2.32.3
python-dotenv==1.0.1
tqdm==4.66.5
pydub==0.25.1
flask==3.0.0
flask-cors==4.0.0